+++ NEWS.md
@@ -1,6 +1,7 @@
 # assemblerr (development version)
 
 * added `task` option to `render` function allowing to customize estimation and covariance records, e.g., `render(m, tasks = tsk_estimation(method = "foce"))`
+* added `tsk_ouput()` to specify variables that should be added to $TABLE
 
 # assemblerr 0.0.0.9006 
 * RUV variance can be specified in the observation model `obs_additive(~conc, var_add = 2)`